Move zoom/lat/long tracking to MapManager
